Changed method to get axiom and angle from user.
After looking at some L-system examples online, I decided it would be better to ask the user explicitly for one angle to use for all turns and the axiom, should it be different than any of the rules. The rules dictionary now has separate entries for angle and axiom. Tests were updated to reflect this and everything is working again.