Built and tested function to recursively generate commands from an L-system
Wrote the meatiest part of the program. Given an axiom and set of L-system rules, recGenerateLString now recursively writes a flat string containing all of the drawing commands for a curve of a particular size. Tests are all in working order. Moving on to calculating actual points on a grid based on the commands next.